added post

+++ 12MegaBlog/src/components/post-form/PostForm.jsx
@@ -3,9 +3,10 @@ import { useForm } from "react-hook-form";
 import { Button, Input, RTE, Select } from "..";
 import appwriteService from "../../appwrite/configuration";
 import { useNavigate } from "react-router-dom";
-import { useSelector } from "react-redux";
+import { useSelector,useDispatch } from "react-redux";
+import { getPosts } from "../../store/postSlice";
 
-export default function PostForm({ post }) {
+export default function PostForm({ post = null }) {
     const { register, handleSubmit, watch, setValue, control, getValues } = useForm({
         defaultValues: {
             title: post?.title || "",
@@ -19,35 +20,36 @@ export default function PostForm({ post }) {
     const userData = useSelector((state) => state.auth.userData);
 
     const submit = async (data) => {
-        if (post) {
+        if (post!=null) {
             const file = data.image[0] ? await appwriteService.uploadFile(data.image[0]) : null;
-
-            if (file) {
-                appwriteService.deleteFile(post.featuredImage);
+    
+            if (file && post.featuredImage) {
+                await appwriteService.deleteFile(post.featuredImage);
             }
-
+    
             const dbPost = await appwriteService.updatePost(post.$id, {
                 ...data,
                 featuredImage: file ? file.$id : undefined,
             });
-
+    
             if (dbPost) {
                 navigate(`/post/${dbPost.$id}`);
             }
         } else {
             const file = await appwriteService.uploadFile(data.image[0]);
-
+    
             if (file) {
                 const fileId = file.$id;
                 data.featuredImage = fileId;
                 const dbPost = await appwriteService.createPost({ ...data, userId: userData.$id });
-
+    
                 if (dbPost) {
                     navigate(`/post/${dbPost.$id}`);
                 }
             }
         }
     };
+    
 
     const slugTransform = useCallback((value) => {
         if (value && typeof value === "string")
